service/requirements.txt
appnope==0.1.0
decorator==4.0.11
Django==1.11
django-cors-headers==2.0.2
django-guardian==1.4.6
djangorestframework==3.5.4
djangorestframework-jsonapi==2.1.1
inflection==0.3.1
itsdangerous==0.24
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.2.0
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
